feat(docker): add .dockerignore file to exclude unnecessary files from Docker context feat(solution): add EterniaEmu.Server project to the solution for server functionality refactor(NetworkPacketAttribute): change constructor to accept PacketTypeEnum for better type safety refactor(PacketTypeEnum): change underlying type from byte to int for better range feat(network): implement INetworkPacket interface and related classes for packet handling feat(CharacterCreationPacket): create CharacterCreationPacket class for character creation data feat(GenderAndRaceType): add GenderAndRaceType enum for character gender and race representation feat(server): add Dockerfile and project files for EterniaEmu.Server to enable containerization style(server): add initial Program.cs file with a simple console output for testing

+++ src/EterniaEmu.Network/Consts/PacketTypeEnum.cs
@@ -1,6 +1,6 @@
 namespace EterniaEmu.Network.Consts;
 
-public enum PacketTypeEnum : byte
+public enum PacketTypeEnum : int
 {
     CharacterCreation = 0x00,
     Logout = 0x01,
